Crate windows_winmd

A Windows Metadata (winmd) parser

Re-exports

pub use parsed::*;

Modules

parsed
Parsed elements of a winmd file

Structs

File
A Windows Metadata File

TypeReader
A reader of type information from Windows Metadata

Enums

TableIndex
A well-known index of data into the winmd tables array

Type

Traits

Decode

Functions

workspace_windows_dir
Returns the build's .windows directory in the root of the workspace as a PathBuf.